Refactor LK util functions into submodules
